%p with printf길이 반환오류

+++ printf/ft_nbrlen.c
@@ -6,13 +6,13 @@
 /*   By: hyson <hyson@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/28 16:46:02 by hyson             #+#    #+#             */
-/*   Updated: 2021/03/28 16:54:41 by hyson            ###   ########.fr       */
+/*   Updated: 2021/04/03 20:04:27 by hyson            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-int	ft_nbrlen(int nbr, int base)
+int	ft_nbrlen(long long  nbr, int base)
 {
 	int len;
